+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/filter/FilterFactoryContext.java
@@ -6,8 +6,6 @@
 
 package io.kroxylicious.proxy.filter;
 
-import java.util.concurrent.ScheduledExecutorService;
-
 import io.kroxylicious.proxy.plugin.UnknownPluginInstanceException;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -17,17 +15,6 @@ import edu.umd.cs.findbugs.annotations.NonNull;
  * to the FilterFactory when it is creating a new instance of the Filter. see {@link FilterFactory#createFilter(FilterFactoryContext, Object)}
  */
 public interface FilterFactoryContext {
-
-    /**
-     * The event loop that this Filter's channel is assigned to.
-     * Null if the factory is not bound to a channel yet.
-     * Should be safe
-     * to mutate Filter members from this executor.
-     * @return executor, or null
-     * @deprecated use {@link #filterDispatchExecutor()} instead
-     */
-    @Deprecated(since = "0.7.0")
-    ScheduledExecutorService eventLoop();
 
     /**
      * An executor backed by the single Thread responsible for dispatching